Generate html files in a format suitable for Qucs
so that the "online help" of each component can point to the
component help page. Actually Qucs will point to the component model
help, which will be a link to the first component using a particular
model.
Each component in the Component Reference pages has two references
(labels):
- an internal one, used for internal links in the html help
- an external one, used by the Qucs GUI, which referes to the
component model